+++ rmk-config/src/lib.rs
@@ -724,6 +724,8 @@ pub(crate) struct MorseProfile {
 
     /// The time elapsed from the last release of a key is longer than this, it will break the morse pattern (in milliseconds)
     pub gap_timeout: Option<DurationMillis>,
+
+    pub quick_tap_timeout: Option<DurationMillis>,
 }
 
 /// Configurations for tri layer
@@ -846,6 +848,8 @@ pub(crate) struct MorsesConfig {
 
     /// The time elapsed from the last release of a key is longer than this, it will break the morse pattern (in milliseconds)
     pub gap_timeout: Option<DurationMillis>,
+
+    pub quick_tap_timeout: Option<DurationMillis>,
 
     /// these can be used to overrides the defaults given above
     pub profiles: Option<HashMap<String, MorseProfile>>,